차트 데이터 버그 수정 마켓 데이터 연동

+++ CoinGecko/Extension/Double+.swift
@@ -1,5 +1,5 @@
 //
-//  Int+.swift
+//  Double+.swift
 //  CoinGecko
 //
 //  Created by 홍정민 on 9/8/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     var toFormattedString: String {
         return "₩" + self.formatted()
     }
